adjusted wave enemies

+++ src/utils/GameManager.cs
@@ -9,7 +9,7 @@ namespace BulletHell.Utils
     {
         public const int FRAMES_PER_SECOND = 60;
         public const int TICKS_PER_SECOND = 60;
-        public const float TICK_STEP = 1f / TICKS_PER_SECOND;
+        public const double TICK_STEP = 1f / TICKS_PER_SECOND;
 
         private const double TIME_SCALE_MIN = 0;
         private const double TIME_SCALE_MAX = 5;
